Moved API schemas to app.api.schemas
