+++ application/backend/app/schemas/__init__.py
@@ -8,7 +8,6 @@ from app.schemas.model import Model, ModelFormat
 from app.schemas.model_architecture import ModelArchitectures
 from app.schemas.pipeline import DataCollectionPolicy, PipelineStatus, PipelineView
 from app.schemas.project import ProjectCreate, ProjectUpdateName, ProjectView
-from app.schemas.sink import SinkCreate, SinkView
 from app.schemas.source import DisconnectedSourceConfig, Source, SourceCreate, SourceType
 from app.schemas.training_configuration import TrainingConfiguration
 
@@ -31,8 +30,6 @@ __all__ = [
     "ProjectCreate",
     "ProjectUpdateName",
     "ProjectView",
-    "SinkCreate",
-    "SinkView",
     "Source",
     "SourceCreate",
     "SourceType",
